backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
python-dotenv==1.0.1
Pillow==10.4.0
python-multipart==0.0.9
